Extract `JetpackScanHistoryStats` in dedicated file
